+++ src/test/26-manipulating-views.ts
@@ -1,6 +1,7 @@
 import { expect } from "chai";
 import { Database } from "../databases.js";
 import { ArangoSearchViewProperties, View } from "../views.js";
+import { fetchArangoVersionCode } from "./_arango-server-version.js";
 import { config } from "./_config.js";
 import {
   clusterIntegrationTimeoutMs,
@@ -11,20 +12,22 @@ import {
 // NOTE These tests will not reliably work in a cluster.
 const describeNLB =
   config.loadBalancingStrategy === "ROUND_ROBIN" ? describe.skip : describe;
-const it312 = config.arangoVersion! >= 31200 ? it : it.skip;
 
 describe("Manipulating views", function () {
   this.timeout(clusterIntegrationTimeoutMs);
   const name = `testdb_${Date.now()}`;
   let system: Database, db: Database;
   let view: View;
+  let arangoVersionCode: number;
   before(async () => {
     system = new Database(config);
     if (Array.isArray(config.url) && config.loadBalancingStrategy !== "NONE")
       await system.acquireHostList();
     db = await system.createDatabase(name);
     await waitForNewDatabase(db);
+    arangoVersionCode = await fetchArangoVersionCode(db);
   });
+
   after(async () => {
     try {
       await system.dropDatabase(name);
@@ -73,29 +76,34 @@ describe("Manipulating views", function () {
       expect(properties.consolidationIntervalMsec).to.equal(45000);
       expect(properties.commitIntervalMsec).to.equal(30000);
     });
-    it312(
-      "should support new consolidation policy options (maxSkewThreshold, minDeletionRatio)",
-      async () => {
-        const properties = (await view.updateProperties({
-          consolidationPolicy: {
-            type: "tier",
-            maxSkewThreshold: 0.5,
-            minDeletionRatio: 0.6,
-          },
-        })) as ArangoSearchViewProperties;
-        expect(properties.consolidationPolicy).to.have.property("type", "tier");
-        if (properties.consolidationPolicy.type === "tier") {
-          expect(properties.consolidationPolicy).to.have.property(
-            "maxSkewThreshold",
-            0.5,
-          );
-          expect(properties.consolidationPolicy).to.have.property(
-            "minDeletionRatio",
-            0.6,
-          );
-        }
-      },
-    );
+    describe("new consolidation policy options", function () {
+      before(function () {
+        if (arangoVersionCode < 31200) this.skip();
+      });
+      it(
+        "should support maxSkewThreshold and minDeletionRatio",
+        async () => {
+          const properties = (await view.updateProperties({
+            consolidationPolicy: {
+              type: "tier",
+              maxSkewThreshold: 0.5,
+              minDeletionRatio: 0.6,
+            },
+          })) as ArangoSearchViewProperties;
+          expect(properties.consolidationPolicy).to.have.property("type", "tier");
+          if (properties.consolidationPolicy.type === "tier") {
+            expect(properties.consolidationPolicy).to.have.property(
+              "maxSkewThreshold",
+              0.5,
+            );
+            expect(properties.consolidationPolicy).to.have.property(
+              "minDeletionRatio",
+              0.6,
+            );
+          }
+        },
+      );
+    });
   });
   describe("view.replaceProperties", () => {
     it("should overwrite properties", async () => {